Youtube_Downloader-main/requirements.txt
pytube3==9.6.4
requests==2.26.0
Pillow==8.3.2
PySimpleGUI==4.49.0
pytube==11.0.1
